rt0: Convert uses of u32 to usize
u32 is not large enough to hold pointers on 64-bit architectures.

BUG=none
TEST=make

(cherry picked from commit 57a9b2a38400913e42bae4efa65e8977eafc717b)

+++ chips/nrf52/src/crt1.rs
@@ -14,11 +14,11 @@ use tock_rt0;
 
 extern "C" {
     // Symbols defined in the linker file
-    static mut _erelocate: u32;
-    static mut _etext: u32;
-    static mut _ezero: u32;
-    static mut _srelocate: u32;
-    static mut _szero: u32;
+    static mut _erelocate: usize;
+    static mut _etext: usize;
+    static mut _ezero: usize;
+    static mut _srelocate: usize;
+    static mut _szero: usize;
     fn reset_handler();
 
     // _estack is not really a function, but it makes the types work
